+++ README.md
@@ -2,6 +2,9 @@
 
 Simple react app samples to learningn the foundations of React
 
+## Components
+- [Social Cards](./src/pages/social-cards/README.md)
+
 # Build
 
-Project created with CRA, use the same API.
+Project created with [CRA](https://facebook.github.io/create-react-app/docs/getting-started), use the same API.
