+++ src/pages/shared_component/spinner/spinner.js
@@ -1,0 +1,18 @@
+import React from 'react'
+// import PropTypes from 'prop-types'
+import './spinner.scss';
+
+function Spinner(props) {
+  return (
+    <div className="spinner spinner--container">
+      <i class="fas fa-spinner" />
+    </div>
+  )
+}
+
+// spinner.propTypes = {
+
+// }
+
+export default Spinner
+
