+++ src/pages/shared_component/spinner/spinner.scss
@@ -1,0 +1,17 @@
+.spinner {
+  display: block;
+  position: relative;
+  margin: auto;
+  font-size: 1.6em;
+  width: fit-content;
+
+  i {
+    -webkit-animation:spin 1s linear infinite;
+    -moz-animation:spin 1s linear infinite;
+    animation:spin 1s linear infinite;
+  }
+
+  @-moz-keyframes spin { 100% { -moz-transform: rotate(360deg); } }
+  @-webkit-keyframes spin { 100% { -webkit-transform: rotate(360deg); } }
+  @keyframes spin { 100% { -webkit-transform: rotate(360deg); transform:rotate(360deg); } }
+}
