+++ src/pages/social-cards/components/card-action/card-action.scss
