+++ src/pages/social-cards/components/external-source/external-source.js
