+++ src/pages/social-cards/components/social-card/social-card.scss
