+++ src/pages/social-cards/social-cards-page.scss
@@ -1,0 +1,18 @@
+.stories--container {
+  margin: 20px auto;
+  border: 1px solid #e1e8ed;
+  width: fit-content;
+  
+  .social-card {
+    border-bottom: 1px solid #e1e8ed;
+   
+    &:hover {
+      background: #f5f8fa;
+    }
+  }
+
+}
+.spinner--container {
+  margin-top:  40px;
+}
+
